Rename package application to allu
