+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/Configuration.kt
@@ -1,9 +1,9 @@
 package fi.hel.haitaton.hanke
 
-import fi.hel.haitaton.hanke.application.AlluProperties
-import fi.hel.haitaton.hanke.application.ApplicationRepository
-import fi.hel.haitaton.hanke.application.ApplicationService
-import fi.hel.haitaton.hanke.application.CableReportServiceAllu
+import fi.hel.haitaton.hanke.allu.AlluProperties
+import fi.hel.haitaton.hanke.allu.ApplicationRepository
+import fi.hel.haitaton.hanke.allu.ApplicationService
+import fi.hel.haitaton.hanke.allu.CableReportServiceAllu
 import fi.hel.haitaton.hanke.geometria.HankeGeometriatDao
 import fi.hel.haitaton.hanke.geometria.HankeGeometriatDaoImpl
 import fi.hel.haitaton.hanke.geometria.HankeGeometriatService
