+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/allu/CableReportApplication.kt
@@ -1,4 +1,4 @@
-package fi.hel.haitaton.hanke.application
+package fi.hel.haitaton.hanke.allu
 
 import org.geojson.GeometryCollection
 import java.time.ZonedDateTime
