+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/allu/InformationRequest.kt
@@ -1,4 +1,4 @@
-package fi.hel.haitaton.hanke.application
+package fi.hel.haitaton.hanke.allu
 
 data class InformationRequest(
         val applicationId: Int,
